fix paging issue with data adapter
use long intern, but it's useless right now...

+++ MComponents/MGrid/GridState/MGridState.cs
@@ -6,7 +6,7 @@ namespace MComponents.MGrid
 {
     public class MGridState
     {
-        public int? Page { get; set; }
+        public long? Page { get; set; }
 
         public int? PageSize { get; set; }
